Move a lot of functions into subfolders:
- This makes it easier to navigate the project as it makes files shorter
