+++ internal/clear_test.go
@@ -1,4 +1,4 @@
-package main
+package internal
 
 import (
 	"reflect"
